+++ tests/worker/test_shell.py
@@ -4,7 +4,7 @@ from typing import Any, Callable, Optional, Type
 
 import pytest
 
-from drudgeyer.tools.queue import BaseQueueModel
+from drudgeyer.job_scheduler.queue import BaseQueueModel
 from drudgeyer.worker.logger import BaseLog
 from drudgeyer.worker.shell import BaseWorker, Worker
